# copilot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookYourCoach API",
        "description": "API REST pour la plateforme de réservation de cours avec coaches équestres ou autres sports. Cette API permet la gestion complète des utilisateurs (admin/enseignants/élèves), des réservations, des paiements et de la facturation.",
        "contact": {
            "name": "BookYourCoach Support",
            "url": "https://bookyourcoach.com",
            "email": "support@bookyourcoach.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur de développement BookYourCoach"
        }
    ],
    "paths": {
...
        "/api/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Liste des cours",
                "description": "Récupère la liste des cours selon le rôle de l'utilisateur",
                "operationId": "getLessons",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Date de début de période",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Date de fin de période",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des cours récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lesson"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Créer un nouveau cours",
                "description": "Crée une nouvelle réservation de cours",
                "operationId": "createLesson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teacher_id",
                                    "course_type_id",
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-08-15 14:00:00"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Premier cours de dressage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lesson"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cours créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/students/{id}/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Cours d'un étudiant",
                "description": "Récupère les cours d'un étudiant spécifique",
                "operationId": "getStudentLessons",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étudiant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cours de l'étudiant récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lesson"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Étudiant non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lesson": {
                "title": "Lesson",
                "description": "Cours/leçon réservé sur la plateforme",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du cours",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "teacher_id": {
                        "description": "ID de l'enseignant",
                        "type": "integer",
                        "example": 1
                    },
                    "student_id": {
                        "description": "ID de l'élève",
                        "type": "integer",
                        "example": 1
                    },
                    "course_type_id": {
                        "description": "ID du type de cours",
                        "type": "integer",
                        "example": 1
                    },
                    "location_id": {
                        "description": "ID du lieu",
                        "type": "integer",
                        "example": 1
                    },
                    "scheduled_at": {
                        "description": "Date et heure prévues",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-08-15 14:00:00"
                    },
                    "duration": {
                        "description": "Durée en minutes",
                        "type": "integer",
                        "example": 60
                    },
                    "price": {
                        "description": "Prix du cours",
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "completed",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "notes": {
                        "description": "Notes sur le cours",
                        "type": "string",
                        "example": "Premier cours de dressage"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer. Utilisez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}